refactor(image): 모든 PNG를 WebP 변환 및 코드 참조 업데이트
