+++ src/screens/Home/HomeScreen.js
@@ -185,13 +185,13 @@ const HomeScreen = () => {
             <BannerCard
               title="좋은 제휴 아이디어 공유해주실래요?"
               subtitle="원하는 제휴 혜택을 학생회에게 직접 제안해요"
-              imageSource={require('../../../assets/images/home/banner_03.png')}
+              imageSource={require('../../../assets/images/home/banner_03.webp')}
               onPress={() => console.log('제안하기 클릭')}
             />
             <BannerCard
               title="제휴 이용하고 스탬프 받아가세요!"
               subtitle="제휴만 이용해도 혜택이 팡팡"
-              imageSource={require('../../../assets/images/home/banner_04.png')}
+              imageSource={require('../../../assets/images/home/banner_04.webp')}
               onPress={() => navigation.navigate('Stamp')}
             />
           </View>
